Update install instructions
* Make setup.cfg use requirements.txt for install_requires
* List winsdk in the requirements.txt to make setup easier
* Create a build script for MacOS
* Update README instructions

The setup.cfg requirements were out of date with the requirements.txt.
This meant that if you built a binary using the setup.cfg, it would have
an outdated version of librespot and be unable to login. Given that
setuptools and pip can both use properly formatted PEP-508 syntax and
onthespot is not using any pip-specific syntax, this can be unified into
the requirements.txt file for ease.