Added method for cover art (#169)
After some change on the Tidal API it's now needed to get on another endpoint the ID for the artwork and then it is possible to get the link as before. The Tidal API document site isn't update with this change but I found Tidal's Swagger and it has these changes.